feat: add support for QEMU emulators
Adds a enhanced KAS Container Dockerfile for using QEMU inside the KAS
container (for local testing and also CI).
Adds a KAS config for a standard emulator configuration as well as for
a custom QEMU device.
Adds a KConfig fragment for the linux-raspberrypi kernel to enable
virtio, needed to run it with QEMU.

Signed-off-by: Anna-Lena Marx <anna-lena.marx@inovex.de>